refactor: move masternode mode logic to `chainlock/signer.{cpp,h}`
Review with `git log -p -n1 --color-moved=dimmed_zebra`.
